# Simplify only new types to improve performance on incremental analysis