--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 ### Changed
+
+- Performance of incrementally checking types (i.e. in REPL) was improved (#1483).
+
 ### Deprecated
 ### Removed
 ### Fixed