Hello :octocat: 

This is a simple performance improvement for the REPL (which uses incremental analysis). Instead of iterating over all types each time inference is called, we use set difference and only iterate over new ids.

No behavior should change (other than performance)

@josef-widder has a large spec that is causing the REPL to be quite slow. This improves that, but I have one more meaningful change cooking. See benchmarks to loading + evaluating a single expression in his project:

Before this fix:
```
________________________________________________________
Executed in   46.50 secs    fish           external
   usr time   50.78 secs  184.00 micros   50.78 secs
   sys time    0.79 secs   45.00 micros    0.79 secs

```

After this fix:
```
________________________________________________________
Executed in   36.97 secs    fish           external
   usr time   41.09 secs  175.00 micros   41.09 secs
   sys time    0.66 secs   28.00 micros    0.66 secs
```

<!-- Please ensure that your PR includes the following, as needed -->

- [-] Tests added for any new code
- [-] Documentation added for any new functionality
- [x] Entries added to the respective `CHANGELOG.md` for any new functionality
- [-] Feature table on [`README.md`](../README.md#roadmap) updated for any listed functionality

<!--
Some common CI checks and how to fix them (if failing):
- The formatting in all files is consistent with the project's style.
   - Run `npm run format` to automatically format all files.
- The `examples/README.md` file contains all Quint files in `examples/`
  and correctly lists their ability to go through pipeline stages.
   - Run `make examples` to automatically regenerate this file locally.
- The assets in `quint/testFixture` and `doc/builtin.md` are consistent.
   - Run `npm run generate` to automatically update these files locally.
-->